add another vignette to show how dandelionR can interface with scRepertoire (#37)
* Create fromSceAndVDJ.Rmd

* add scRepertoire

* Update DESCRIPTION

* Update vignette.yml

* fix the bugs

* update notebooks

* Update _pkgdown.yml

* fix option and update docs

* oops missing commas

* make documents fit the line width

* missing reference tab

* modify message

* fix some typos, add citation and rename the vignette names

* add jiawei's changes

* modify parameter filter_pattern

Because character "None" will introduced by splitTCR, so the filter_pattern = ",|None|No_contig" is suitable for both dandelion demo and the scRepetoire combined data. I think it is better to let user decide whether to filter the unmapped data rather than choose the filter_pattern

* Update setupVdjPseudobulk.R

At the end of the main TCR extraction process, the column to be filtered will be generated automatically. What the user needs to specify is whether to filter by this gene, not the column prefix.

* debugs

* update document

* debugs in CITATION

* Update CITATION

* rename files to get rid of warnings

* fix a few more warnings/notes

* rename splitTCR and associated function names and removing some notes

* ok fix the vapply issue

---------

Co-authored-by: Zewen Kelvin Tuong <z.tuong@uq.edu.au>